backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
alembic==1.15.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
boto3==1.38.11
pillow==11.2.1
weasyprint==63.1
jinja2==3.1.6
